+++ libpulse/__init__.py
@@ -1,1 +1,0 @@
-"""An asyncio interface to the Pulseaudio library."""