+++ pa_dlna/__init__.py
@@ -10,7 +10,3 @@ if _version < MIN_PYTHON_VERSION:
     print(f'error: the python version must be at least'
           f' {MIN_PYTHON_VERSION}', file=sys.stderr)
     sys.exit(1)
-
-# Temporary hack to enable importing libpulse at top level.
-from os.path import join, dirname
-sys.path.append(join(dirname(__file__), '..'))